+++ src/frontend/apps/impress/src/features/docs/doc-table-content/components/TableContent.tsx
@@ -1,88 +1,24 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { Box, BoxButton, Text } from '@/components';
-import { Panel } from '@/components/Panel';
 import { useDocStore } from '@/features/docs/doc-editor';
 import { Doc } from '@/features/docs/doc-management';
 
-import { useDocTableContentStore } from '../stores';
+import { HeadingBlock } from '../types';
 
 import { Heading } from './Heading';
 
-const recursiveTextContent = (content: HeadingBlock['content']): string => {
-  if (!content) {
-    return '';
-  }
-
-  return content.reduce((acc, content) => {
-    if (content.type === 'text') {
-      return acc + content.text;
-    } else if (content.type === 'link') {
-      return acc + recursiveTextContent(content.content);
-    }
-
-    return acc;
-  }, '');
-};
-
-type HeadingBlock = {
-  id: string;
-  type: string;
-  text: string;
-  content: HeadingBlock[];
-  contentText: string;
-  props: {
-    level: number;
-  };
-};
-
 interface TableContentProps {
   doc: Doc;
+  headings: HeadingBlock[];
 }
 
-export const TableContent = ({ doc }: TableContentProps) => {
+export const TableContent = ({ doc, headings }: TableContentProps) => {
   const { docsStore } = useDocStore();
   const { t } = useTranslation();
-
   const editor = docsStore?.[doc.id]?.editor;
-  const headingFiltering = useCallback(
-    () =>
-      editor?.document
-        .filter((block) => block.type === 'heading')
-        .map((block) => ({
-          ...block,
-          contentText: recursiveTextContent(
-            block.content as unknown as HeadingBlock['content'],
-          ),
-        })) as unknown as HeadingBlock[],
-    [editor?.document],
-  );
-
-  const [headings, setHeadings] = useState<HeadingBlock[]>();
-  const { setIsPanelTableContentOpen, isPanelTableContentOpen } =
-    useDocTableContentStore();
-  const [hasBeenClose, setHasBeenClose] = useState(false);
-  const setClosePanel = () => {
-    setHasBeenClose(true);
-    setIsPanelTableContentOpen(false);
-  };
-
   const [headingIdHighlight, setHeadingIdHighlight] = useState<string>();
-
-  // Open the panel if there are more than 1 heading
-  useEffect(() => {
-    if (headings?.length && headings.length > 1 && !hasBeenClose) {
-      setIsPanelTableContentOpen(true);
-    }
-  }, [setIsPanelTableContentOpen, headings, hasBeenClose]);
-
-  // Close the panel unmount
-  useEffect(() => {
-    return () => {
-      setIsPanelTableContentOpen(false);
-    };
-  }, [setIsPanelTableContentOpen]);
 
   // To highlight the first heading in the viewport
   useEffect(() => {
@@ -130,68 +66,57 @@ export const TableContent = ({ doc }: TableContentProps) => {
     return null;
   }
 
-  // Update the headings when the editor content changes
-  editor?.onEditorContentChange(() => {
-    setHeadings(headingFiltering());
-  });
-
-  if (!isPanelTableContentOpen) {
-    return null;
-  }
-
   return (
-    <Panel setIsPanelOpen={setClosePanel}>
-      <Box $padding="small" $maxHeight="95%">
-        <Box $overflow="auto">
-          {headings?.map((heading) => (
-            <Heading
-              editor={editor}
-              headingId={heading.id}
-              level={heading.props.level}
-              text={heading.contentText}
-              key={heading.id}
-              isHighlight={headingIdHighlight === heading.id}
-            />
-          ))}
-        </Box>
-        <Box
-          $height="1px"
-          $width="auto"
-          $background="#e5e5e5"
-          $margin={{ vertical: 'small' }}
-          $css="flex: none;"
-        />
-        <BoxButton
-          onClick={() => {
-            editor.focus();
-            document.querySelector(`.bn-editor`)?.scrollIntoView({
+    <Box $padding={{ all: 'small', right: 'none' }} $maxHeight="95%">
+      <Box $overflow="auto">
+        {headings?.map((heading) => (
+          <Heading
+            editor={editor}
+            headingId={heading.id}
+            level={heading.props.level}
+            text={heading.contentText}
+            key={heading.id}
+            isHighlight={headingIdHighlight === heading.id}
+          />
+        ))}
+      </Box>
+      <Box
+        $height="1px"
+        $width="auto"
+        $background="#e5e5e5"
+        $margin={{ vertical: 'small' }}
+        $css="flex: none;"
+      />
+      <BoxButton
+        onClick={() => {
+          editor.focus();
+          document.querySelector(`.bn-editor`)?.scrollIntoView({
+            behavior: 'smooth',
+            block: 'start',
+          });
+        }}
+      >
+        <Text $theme="primary" $padding={{ vertical: 'xtiny' }}>
+          {t('Back to top')}
+        </Text>
+      </BoxButton>
+      <BoxButton
+        onClick={() => {
+          editor.focus();
+          document
+            .querySelector(
+              `.bn-editor > .bn-block-group > .bn-block-outer:last-child`,
+            )
+            ?.scrollIntoView({
               behavior: 'smooth',
               block: 'start',
             });
-          }}
-        >
-          <Text $theme="primary" $padding={{ vertical: 'xtiny' }}>
-            {t('Back to top')}
-          </Text>
-        </BoxButton>
-        <BoxButton
-          onClick={() => {
-            editor.focus();
-            document
-              .querySelector(
-                `.bn-editor > .bn-block-group > .bn-block-outer:last-child`,
-              )
-              ?.scrollIntoView({
-                behavior: 'smooth',
-                block: 'start',
-              });
-          }}
-        >
-          <Text $theme="primary" $padding={{ vertical: 'xtiny' }}>
-            {t('Go to bottom')}
-          </Text>
-        </BoxButton>
-      </Box>
-    </Panel>
+        }}
+      >
+        <Text $theme="primary" $padding={{ vertical: 'xtiny' }}>
+          {t('Go to bottom')}
+        </Text>
+      </BoxButton>
+    </Box>
   );
 };