+++ src/frontend/apps/impress/src/features/docs/doc-versioning/components/ModalVersion.tsx
@@ -90,6 +90,8 @@ export const ModalVersion = ({
               id: docId,
               content: newDoc,
             });
+
+            onClose();
           }}
         >
           {t('Restore')}