+++ src/frontend/apps/impress/src/features/docs/doc-versioning/index.ts
@@ -1,4 +1,3 @@
 export * from './api';
 export * from './components';
-export * from './stores';
 export * from './types';